!242 merge feature/official-hubs into develop

Created-by: spharx01
Commit-by: spharx01;zhixianzhou
Merged-by: spharx01
Description: # Pull Request

## 变更类型

- [ ] Bug 修复
- [ ] 新功能
- [ ] 重构 / 代码清理
- [ ] 性能优化
- [ ] 文档更新
- [ ] 测试补充

## 相关 Issue

Closes #____ (如有)

## 变更摘要

<!-- 一两句话说明改了什么、为什么改 -->

## 检查清单

提交前请确认以下事项：

### 构建 & 测试

- [ ] 本地构建通过：
  ```bash
  cmake -B ../AgentRT-build -DBUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Release
  cmake --build ../AgentRT-build --parallel 8
  ```
- [ ] 测试通过：`cd ../AgentRT-build && ctest --output-on-failure`
- [ ] 代码风格通过：`clang-format --dry-run --Werror` 对所有变更的 `.c`/`.h` 文件

### 代码质量

- [ ] Commit 遵循 [Conventional Commits](https://www.conventionalcommits.org/) 格式（`feat:` / `fix:` / `refactor:` / `docs:` / `test:` / `chore:`）
- [ ] 分支命名规范：`feature/<描述>` 或 `fix/<描述>`
- [ ] 公共 API 附带 Doxygen 注释
- [ ] 无 `TODO` / `FIXME` / `HACK` 残留标记
- [ ] 无硬编码路径或平台相关代码（跨平台兼容使用 `platform.h` 抽象层）

## 影响范围

<!-- 列出受影响的模块和关键文件 -->

| 模块 | 说明 |
|------|------|
|      |      |

---

> **提示**：CI 会自动运行质量检查和测试。如果 CI 报错，请根据日志修复后重新推送即可。
>
> 如有疑问，欢迎在 PR 中直接提问。


See merge request: openairymax/agentos!243

+++ agentos/atoms/coreloopthree/src/cognition/critique/triple_coordinator.c
@@ -417,7 +417,7 @@ agentos_error_t tc3_coordinator_execute(tc3_coordinator_t *coord, const char *in
                     if (expert_opinion)
                         AGENTOS_FREE(expert_opinion);
                 } else {
-                    AGENTOS_LOG_WARN("tc3_coordinator_execute: S1 expert call failed (exp_err=%d unit_index=%zu)", (int)exp_err, unit.unit_index);
+                    AGENTOS_LOG_WARN("tc3_coordinator_execute: S1 expert call failed (exp_err=%d unit_index=%u)", (int)exp_err, (unsigned int)unit.unit_index);
                 }
                 coord->stats.escalated_units++;
                 break;
@@ -441,7 +441,7 @@ agentos_error_t tc3_coordinator_execute(tc3_coordinator_t *coord, const char *in
             AGENTOS_FREE(correction_prompt);
 
             if (err != AGENTOS_SUCCESS || !corrected) {
-                AGENTOS_LOG_WARN("tc3_coordinator_execute: S2 correction call failed (err=%d corrected=%p attempt=%u unit_index=%zu)", (int)err, (void *)corrected, correction_attempts, unit.unit_index);
+                AGENTOS_LOG_WARN("tc3_coordinator_execute: S2 correction call failed (err=%d corrected=%p attempt=%u unit_index=%u)", (int)err, (void *)corrected, correction_attempts, (unsigned int)unit.unit_index);
                 if (critique)
                     AGENTOS_FREE(critique);
                 break;
